Angularizing
Dipping the toes in the angular waters.  Ditch the static HTML and use
an angular script to build an array of comments, which is iterated by
the view.

All very basic - the angular script is inline in the HTML file and there
is no persistence or ability to add new comments etc.