Using bcrypt (bcryptjs)
The book recommends use of bcrypt, but that has dependencies on
python/openssl/others on Windows.  bcryptjs has no such dependencies, so
used that instead even though it's slower than bcrypt.

The username/password are still stored in a local array instead of Mongo
- that will be next on the list.  The password (literal string "pass")
was hashed outside of the app using bcryptjs and is stored in the array
in hashed form.  It will be hashed on-the-fly inside the app when the
login functionality is implemented